refactor(dev): cycle 10 - extract build.py, setup dev/__init__.py, delete commands/dev.py [skip-gen]
